requirements.txt
pip_upgrader==1.4.15
jupyterlab==4.0.5
black==23.7.0
black[jupyter]==23.7.0
coverage==7.3.1
flake8==6.1.0
mypy==1.5.1
pytest==7.4.2
pytest-cov==4.1.0
langchain==0.0.284
pandas==2.0.3
tiktoken==0.4.0
openai==0.28.0
streamlit==1.26.0
streamlit-chat==0.1.1
python-dotenv==1.0.0
unstructured==0.10.12
chromadb==0.3.29
markdown==3.4.4
pypdf==3.15.5
